Correção de login para AuthController
