+++ src/Controller/Auth/AuthController.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace App\Controller\Auth;
+
+use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Template;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\Security\Http\Authentication\AuthenticationUtils;
+
+/**
+ * Class DefaultController
+ * @package App\Controller\Auth
+ * @Route("/auth", name="auth_")
+ */
+class AuthController extends Controller
+{
+    /**
+     * @Route("/login", name="default_login")
+     * @Template("/admin/default/login.html.twig")
+     * @param Request $request
+     * @param AuthenticationUtils $authUtils
+     * @return array
+     */
+    public function login(Request $request, AuthenticationUtils $authUtils)
+    {
+        $error = $authUtils->getLastAuthenticationError();
+        $lastUsername = $authUtils->getLastUsername();
+        return [
+            'error' => $error,
+            'last_username' => $lastUsername
+        ];
+    }
+
+}
