+++ templates/admin/default/login.html.twig
@@ -12,7 +12,7 @@
         <section class="login_content">
             <br />
             <br />
-            <form  method="post" action="{{ path('admin_default_login') }}">
+            <form  method="post" action="{{ path('auth_default_login') }}">
                 <h1>Área Administrativa</h1>
                 <div>
                     <input type="email" name="email" class="form-control" placeholder="email@dominio.com.br" required />
